Split generic inference conflict diagnostics
